requirements.txt
# Core dependencies
streamlit>=1.28.0
elevenlabs>=1.0.0
google-generativeai>=0.3.0

# Audio processing
pydub>=0.25.1

# FastAPI (for api.py if needed)
fastapi>=0.104.0
pydantic>=2.0.0

# Additional utilities
python-multipart>=0.0.6
